+++ lib/freydis/config.rb
@@ -6,6 +6,7 @@ require 'pathname'
 require 'mods/msg'
 
 module Freydis
+  # Loads/Save config variable from a yaml file
   class Config
     include Msg
     attr_reader :opts
@@ -19,17 +20,14 @@ module Freydis
         disk_is_encrypt: false,
         gpg_recipient: '',
         backup_paths: [],
-        exclude_paths: []
+        exclude_paths: [],
+        restore_at: '/'
       }
     end
 
     def save(opts)
       FileUtils.mkdir_p Pathname.new(@cpath).parent.to_s
-      @opts[:disk] = opts[:disk] || ''
-      @opts[:disk_is_encrypt] = opts[:disk_is_encrypt] || false
-      @opts[:gpg_recipient] = opts[:gpg_recipient] || ''
-      @opts[:backup_paths] = opts[:backup_paths] || []
-      @opts[:exclude_paths] = opts[:exclude_paths] || []
+      load_opts(opts)
       File.write @cpath, YAML.dump(@opts)
       success "Saving options to #{@cpath}..."
     end
@@ -42,6 +40,17 @@ module Freydis
         info "Creating config file #{@cpath}..."
         save
       end
+    end
+
+    private
+
+    def load_opts(args)
+      @opts[:disk] = args[:disk] || ''
+      @opts[:disk_is_encrypt] = args[:disk_is_encrypt] || false
+      @opts[:gpg_recipient] = args[:gpg_recipient] || ''
+      @opts[:backup_paths] = args[:backup_paths] || []
+      @opts[:exclude_paths] = args[:exclude_paths] || []
+      @opts[:restore_at] = args[:restore_at] || '/'
     end
   end
 end