+++ lib/freydis.rb
@@ -9,7 +9,6 @@ require_relative 'freydis/cryptsetup'
 require_relative 'freydis/rsync'
 require_relative 'freydis/error'
 require_relative 'freydis/guard'
-require_relative 'freydis/secrets'
 require_relative 'freydis/main'
 
 module Freydis