+++ lib/freydis/guard.rb
@@ -1,16 +1,16 @@
 # frozen_string_literal: true
 
 module Freydis
+  # Guard control argument/path/input and quit if no valid
   module Guard
     module_function
 
     def disk(name)
-      full_path = "/dev/#{name}"
-      raise Freydis::InvalidDisk, 'No disk, use with -d DISK.' unless name
-      raise Freydis::InvalidDisk, 'No disk, use with -d DISK.' if name == ''
-      raise Freydis::InvalidDisk, 'Bad name #{name}, should match with sd[a-z]' unless name.match(/^sd[a-z]{1}$/)
-      raise Freydis::InvalidDisk, "No disk #{full_path} available." unless File.exist? full_path
-      Freydis::Disk.new(full_path).search_id # return disk(name) by-id
+      raise Freydis::InvalidDisk, 'No disk, use with --disk PATH.' unless name
+      raise Freydis::InvalidDisk, 'No disk, use with --disk PATH.' if name == ''
+      raise Freydis::InvalidDisk, "No disk #{name} available." unless File.exist? name
+
+      name
     rescue Freydis::InvalidDisk => e
       puts "#{e.class} => #{e}"
       exit 1
@@ -23,11 +23,12 @@ module Freydis
       exit 1
     end
 
-    def isLuks(disk)
-      raise Freydis::InvalidLuksDev, "No disk." unless disk
+    def luks?(disk)
+      raise Freydis::InvalidLuksDev, 'No disk, use with --disk PATH.' unless disk
       raise Freydis::InvalidLuksDev, "#{disk} does not exist." unless File.exist? disk
+
       sudo = Process.uid != 0 ? 'sudo' : ''
-      if !system(sudo, 'cryptsetup', 'isLuks', disk)
+      unless system(sudo, 'cryptsetup', 'isLuks', disk)
         raise Freydis::InvalidLuksDev, "#{disk} is not valid Luks device."
       end
     rescue Freydis::InvalidLuksDev => e
@@ -35,15 +36,16 @@ module Freydis
       exit 1
     end
 
-    def path?(p)
-      raise Freydis::InvalidPath, "#{p} does not exist." unless File.exist? p
+    def path?(path)
+      raise Freydis::InvalidPath, "#{path} does not exist." unless File.exist? path
     rescue Freydis::InvalidPath => e
       puts "#{e.class} => #{e}"
       exit 1
     end
 
     def gpg(recipient)
-      raise Freydis::GPG, "No recipient, use --gpg-recipient NAME" unless recipient
+      raise Freydis::GPG, 'No recipient, use --gpg-recipient NAME' unless recipient
+
       recipient
     rescue Freydis::GPG => e
       puts "#{e.class} => #{e}"