+++ lib/freydis/rsync.rb
@@ -3,60 +3,43 @@
 require 'mods/exec'
 
 module Freydis
+  # Interact with rsync from unix
   class Rsync
     include Exec
 
-    def initialize
+    def initialize(opts)
       @workdir = '/mnt/freydis/backup'
-      @exclude_paths = %w[
-        "/dev/*"
-        "/proc/*"
-        "/sys/*"
-        "/tmp/*"
-        "/run/*"
-        "/mnt/*"
-        "/media/*"
-        "/var/lib/dhcpcd/*"
-        "*/.gvfs"
-        "*/.vim/*"
-        "*/.weechat/*"
-        "*/.thumbnails/*"
-        "*/.oh-my-zsh/*"
-        "*/.cache/*"
-        "*/.emacs.d/*"
-        "*/.local/share/*"
-        "*/.Xauthority"
-        "*/.xsession-errors"
-        "*/.quickemu/*"
-        "*/.config/BraveSoftware/*"
-        "*/.config/Min/*"
-        "*/.config/emacs"
-        "*/build/*"
-        "*/tmp/*"
-        "*/.npm"
-        "*/.history"
-        "*lost+found"
-      ]
-      #@opts = '-aAXHvR'
-      @opts = '-aAXHv --relative'
-      #@opts = '-aAXHvRx'
+      @exclude_paths = %w['/dev/*' '/proc/*'
+                          '/sys/*' '/tmp/*'
+                          '/run/*' '/mnt/*'
+                          '/media/*' '/home/*/.gvfs'
+                          '/var/lib/dhcpcd/*' '*lost+found']
+      @backup = opts[:backup_paths] || []
+      @user_excludes = opts[:exclude_paths] || []
+      @restore_at = opts[:restore_at] || '/'
+      @opts = '-aAXHv --relative -hh'
     end
 
     def backup
-      DiskLuks.open
+      raise 'Nothing to backup, use --paths-add PATH' if @backup == []
+
       mkdir @workdir
-      exil = @exclude_paths * ','
-      save = OPTIONS[:backup_paths] * ' '
-      @opts += ' --delete'
+      exil = combine_exclude
+      save = @backup * ' '
+      @opts += ' --delete --recursive'
       x "rsync #{@opts} --exclude={#{exil}} #{save} #{@workdir}"
       puts "Saved path #{save}"
-      DiskLuks.close
     end
 
     def restore
-      DiskLuks.open
-      x "rsync #{@opts} #{@workdir} /"
-      DiskLuks.close
+      x "rsync #{@opts} #{@workdir} #{@restore_at}"
+    end
+
+    private
+
+    def combine_exclude
+      new_array = @exclude_paths << @user_excludes
+      new_array.flatten! * ','
     end
   end
 end