+++ CHANGELOG.md
@@ -1,6 +1,6 @@
-* Option `--disk PATH` now require the full path, you can add yourself by-id, by-uuid, etc...
-* Remove cli option `--path-del`, edit the config file if need to do this.
-* User can customize what paths to include or exclude `--exclude-paths DIR`.
+* New option `--restore-at PATH` to change the default `/`.
+* Option `--disk PATH` now require the full path (e.g: `/dev/sdx`), you can add yourself by-id, by-uuid, etc...
+* User can customize what paths to include `--paths-add PATHS` or exclude `--paths-del PATHS`.
 * Encrypted disk is no more default, use `--disk-encrypt` if need.
 
 ## 0.2.0, release 2023/10/23