+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* New workflow (github gem)
+
 ## 0.2.0, release 2023/10/23
 * Define constant OPTIONS and ACTIONS
 * Correct rsync options